Shifted API expections
