+++ core/exceptions/__init__.py
@@ -1,0 +1,2 @@
+from .ups import UPSAPIError
+from .usps import USPSAPIError
